refactor(cost): extract processing state repository
