+++ src/services/portfolio_transaction_processing_service/app/infrastructure/cost_basis/__init__.py
@@ -4,12 +4,18 @@ from .corporate_action_reconciliation_repository import (
     SqlAlchemyCorporateActionReconciliationRepository,
 )
 from .fx_rate_repository import SqlAlchemyCostBasisFxRateRepository
+from .processing_state_repository import (
+    SqlAlchemyCostBasisProcessingStateRepository,
+    cost_basis_processing_lock_key,
+)
 from .reference_data_repository import SqlAlchemyCostBasisReferenceDataRepository
 from .staged_effects import StagedCostEffects
 
 __all__ = [
     "SqlAlchemyCorporateActionReconciliationRepository",
     "SqlAlchemyCostBasisFxRateRepository",
+    "SqlAlchemyCostBasisProcessingStateRepository",
     "SqlAlchemyCostBasisReferenceDataRepository",
     "StagedCostEffects",
+    "cost_basis_processing_lock_key",
 ]
